Изменение : Исправил проблему с сохранением настроек во внешнем фильтре видео-декодера, пока как временное решение просто взяли другой тип переменной - "с запасом" :)
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@3562 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/transform/MPCVideoDec/IMPCVideoDec.h
@@ -79,8 +79,8 @@ public IUnknown {
 
 	STDMETHOD_(GUID*, GetDXVADecoderGuid()) = 0;
 
-	STDMETHOD(SetActiveCodecs(DWORD nValue)) = 0;
-	STDMETHOD_(DWORD, GetActiveCodecs()) = 0;
+	STDMETHOD(SetActiveCodecs(ULONGLONG nValue)) = 0;
+	STDMETHOD_(ULONGLONG, GetActiveCodecs()) = 0;
 
 	STDMETHOD_(LPCTSTR, GetVideoCardDescription()) = 0;
